change post id type

+++ backend/curator/src/main/java/com/ssafy/curator/repository/post/PostImageRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface PostImageRepository extends JpaRepository<PostImageEntity, Integer> {
-    List<PostImageEntity> findByPostId(int id);
+    List<PostImageEntity> findByPostId(Long id);
 }
